introduced universal shipment label generator integration
