+++ requirements.server.dev.txt
@@ -4,7 +4,7 @@ Django==3.2.9
 # -e ./schemas/dhl_poland
 
 # Carrier Extenstions in Development
-# -e ./sdk/core
+-e ./sdk/core
 -e ./sdk/extensions/canadapost
 # -e ./sdk/extensions/canpar
 # -e ./sdk/extensions/dhl_express
