+++ sdk/core/purplship/universal/mappers/__init__.py
@@ -1,2 +1,5 @@
-from purplship.universal.mappers.settings import Settings
-from purplship.universal.mappers.proxy import Proxy
+from purplship.universal.mappers.settings import (
+    RatingMixinSettings,
+    ShippingMixinSettings,
+)
+from purplship.universal.mappers.proxy import ShippingMixinProxy, RatingMixinProxy
