+++ sdk/core/purplship/universal/providers/__init__.py
@@ -1,1 +1,5 @@
 from purplship.universal.providers.rate import parse_rate_response, rate_request
+from purplship.universal.providers.shipment import (
+    parse_shipment_response,
+    shipment_request,
+)
